+++ stwo_cairo_prover/crates/prover/src/components/range_check_builtin/mod.rs
@@ -1,2 +1,2 @@
 pub mod component;
-pub mod simd_trace;
+pub mod prover;
